refactor config fragments to BaseGosPkgStateConfigFragment
